feat(tui): simplify footer transition to crossfade
