+++ packages/tui/src/feature-plugins/sidebar/footer.tsx
@@ -1,13 +1,13 @@
 import { Plugin } from "@opencode-ai/plugin/tui"
 import { createMemo } from "solid-js"
-import { DissolveFilePath } from "../../ui/dissolve-file-path"
+import { FadeFilePath } from "../../ui/fade-file-path"
 
 function View(props: { context: Plugin.Context }) {
   const directory = createMemo(() =>
     props.context.location ? props.context.ui.format.path(props.context.location.directory) : undefined,
   )
   return (
-    <DissolveFilePath
+    <FadeFilePath
       value={directory()}
       maxWidth={38}
       fg={props.context.theme.text.subdued}
